// Pessoa/src/app/pessoas/model/pessoa.ts
// export interface Course { //ajuda durante o desenvolvimento
//   _id: string;
//   name: string;
//   category: string;
// }

export interface Pessoa { //ajuda durante o desenvolvimento
  id: number;
  nome: string;
  cpf: string;
  idade: number;
  peso: number;
  bairro: string;
  cidade: string;
  estado: string;
  dataNasc: string;
}